backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
supabase==2.3.0
python-dotenv==1.0.0
google-generativeai==0.3.1
twilio==8.5.0
websockets==11.0.3
httpx==0.24.1
python-multipart==0.0.6
feedparser==6.0.10
yfinance==0.2.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
tenacity==8.2.3
elevenlabs==1.3.0
aiohttp==3.9.5  